Give every external function a prototype
-Wmissing-prototypes reported 640 functions. 494 used only in their own
file are now static. Functions called from other files, including the
test hooks, got one prototype in the header that matches the defining
source file, and every local copy in other source and test files was
removed. The db.c test hooks were declared inside db.h's __DB_C__
section, which db.c itself compiles out.

Obsolete code nothing calls was deleted: do_drink_old, do_eat_old,
load_contextual_hints_legacy, test_tokenize, the retired spatial audio
test and meteor emitters with the two helpers only they used, the
unregistered BoundsCheckingSuite, and dead copies of parseadminlevel
and parselast in rebuildMailIndex. Other never-called functions keep a
prototype and are listed in the plan as dead-code candidates.

damage_handling now exists only for its LUMINARI_CUTEST hook, and two
treasure helpers only under USE_NEW_CRAFTING_SYSTEM like their caller.

missing-prototypes falls from 640 to 2 on GCC 16.2 and Clang 22.1.8.

+++ src/ai_service.c
@@ -2097,7 +2097,7 @@ static char *make_ollama_request(const char *prompt)
     {
       log("AI Service: ERROR - Ollama CURL error: %s", curl_easy_strerror(res));
     }
-    AI_DEBUG("Ollama CURL error code: %d", res);
+    AI_DEBUG("Ollama CURL error code: %d", (int)res);
   }
 
   /* Cleanup */
